==> 05_descriptive_analysis.Rmd <==
---
title: "Descriptive_analysis"
author: "KBazemore"
date: "2024-12-02"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

knitr::opts_knit$set(root.dir = "/s3buckets/adamnaj-lab-psom-s3-bucket-01/bazemork/EPID6210_Final")
```

libraries
```{r}
library(dplyr)
library(cardx)
library(bigsnpr)

library(reshape2)
library(ggplot2)
library(dplyr)
library(corrr)

#url <- "https://cran.r-project.org/src/contrib/Archive/qlspack/qlspack_2.2.tar.gz"
#pkgFile <- "qlspack_2.2.tar.gz"
#download.file(url = url, destfile = pkgFile)
#install.packages(pkgs=pkgFile, type="source", repos=NULL)
library(qlspack)
```

Read in data
```{r}
ukbbdat <- readRDS("data/Prepped_UKBB_B.rds")
```

Demographics table
```{r}
ukbbdat$sex <- as.factor(ukbbdat$sex)
ukbbdat$AD <- as.factor(ukbbdat$A)

ukbbdat %>% 
  dplyr::select(AD, participant_age_recriutment, sex, e4_count, pop_cat, score.norm) %>%
  dplyr::summarise(n.obs = n(),
                   Mean.PRS = format(mean(score.norm, na.rm = T), scientific = T, digits = 3),
                   SD.PRS = round(sd(score.norm, na.rm = T), 2),
                   Mean.Age = round(mean(participant_age_recriutment, na.rm = T), 2),
                   SD.Age = round(sd(participant_age_recriutment, na.rm = T), 2),
                   n.AD = sum(AD == 1, na.rm = T),
                   per.AD = round((sum(AD == 1, na.rm = T)/n())*100, 2),
                   per.Male = round((sum(sex == 1, na.rm = T)/n())*100, 2),   #double check that sex == 1 means male
                   per.e4Carrier = round((sum(e4_count == 1 | e4_count ==2, na.rm = T)/n())*100, 2))
```
Fluid memory summary
```{r}
ukbbdat <- ukbbdat %>% dplyr::select("eid", "e4_count", "AD", "participant_age_recriutment", "sex", "pc1", "pc2", "pc3", "pc4", "M0", "M1", "M2", "M3", "score", "score.norm")

#Reshape into long format
ukbb_long <- ukbbdat %>% reshape2::melt(
  id.vars = c("eid", "e4_count", "AD", "participant_age_recriutment", "sex", "pc1", "pc2", "pc3", "pc4", "score", "score.norm"),
  variable.name = c("time")) %>%
    dplyr::arrange(eid, time)

#descriptive table
ukbb_long %>% 
  dplyr::group_by(time) %>% 
  dplyr::summarise(n.obs = sum(!is.na(value)),
                   Mean.FI=mean(value, na.rm = T),
                   SD = sd(value, na.rm = T),
                   Median.FI = median(value, na.rm = T))
```

Spaghetti plot of fluid intelligence over time 
```{r}
suppressWarnings(ggplot(ukbb_long) + 
  geom_line(aes(x = time, y = value, group = eid, color = factor(eid)), show.legend = F) +
  stat_summary(aes(x = time, y = value), geom="line", fun=mean, color="red", linewidth = 1.2) +
  xlab("Measurement Visit") +
  ylab("Fluid Intelligence Score") +
  labs(title="Individual Fluid Intelligence in UK Biobank Participants over 50") +
  theme_bw())
```
Correlation of fluid intelligence scores over time
```{r}
ukbbdat %>% select(M0, M1, M2, M3) %>%
  correlate(., y = NULL, use ='pairwise.complete.obs', method = 'pearson')
```
